improvement(pulumi): improve preview output
We now only show the output of `pulumi preview` when the generated plan
is not a no-op. This makes the output much easier to parse visually
during a manual review process.

We now also copy plans that aren't no-ops to a subdirectory of the preview
directory for easier review and deployment.